refactor: modularize lofi renderer
